+++ src/main/java/com/doodle/poll/domain/PollRepo.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PollRepo extends CrudRepository<Poll, Integer> {
+public interface PollRepo extends CrudRepository<Poll, Long> {
 
 	@Query("select p from Poll p where p.title like %:title%")
 	List<Poll> searchByTitle(String title);
